Create multi-page app with recipe, shopping, and home pages
